+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `OtcMaker` contract.
 
+## [1.4.3] - 2023-04-11
+### Added
+- Add `PerpPortal.getSqrtMarketTwapX96()` to return market twap.
+
+### Deprecated
+- `PerpPortal.getSqrtMarkTwapX96(address baseToken, uint32 twapInterval)` will be deprecated at later releases. Suggest to use `PerpPortal.getSqrtMarketTwapX96()` instead.
+
+## [1.4.2] - 2023-04-10
+- Support Curie v2.6.0
+
 ## [1.4.1] - 2022-08-01
 
 ### Removed